first commit to start-sharedPreferens: "move SharedPreferens "
